requirements.txt
anyio==4.12.1
argcomplete==3.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blivet==3.12.1
Brlapi==0.8.7
certifi==2025.7.9
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-356-build/cockpit-356/tmp/wheel/cockpit-356-py3-none-any.whl#sha256=5a84a26ba9e7f611c00c6efb3a3e675039514318856e1355bd01b72f1d86f01c
crypt_r==3.13.1
cupshelpers==1.0
dasbus==1.7
dbus-python==1.4.0
dbus_next==0.2.3
distro==1.9.0
dnf==4.24.0
fedora-third-party==0.10
file-magic==0.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.13.0
idna==3.10
iso639==0.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.69
libcomps==0.1.23
libdnf==0.75.0
libdnf5==5.2.18.0
louis==3.33.0
lxml==6.0.1
nftables==0.1
numpy==2.4.2
olefile==0.47
packaging==25.0
pandas==3.0.1
perf==0.1
pexpect==4.9.0
pid==3.0.4
pillow==11.3.0
ply==3.11
productmd==1.50
progressbar2==4.5.0
ptyprocess==0.7.0
pwquality==1.4.5
PyAudio==0.2.13
pycairo==1.28.0
pycparser==2.22
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.54.5
pyinotify==0.9.6
pykickstart @ file:///builddir/build/BUILD/pykickstart-3.66-build/pykickstart-3.66
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.9.0.post0
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-utils==3.9.1
pyudev==0.24.3
pyxdg==0.28
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.18.12-build/kernel-6.18.12/linux-6.18.12-200.fc43.x86_64/tools/net/ynl
RapidFuzz==3.12.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
requests-file==2.0.0
requests-ftp==0.3.1
rpds-py==0.27.0
rpm==6.0.1
rpmautospec==0.8.3
rpmautospec-core==0.1.5
selinux @ file:///builddir/build/BUILD/libselinux-3.9-build/libselinux-3.9/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.9-build/selinux-3.9/python/sepolicy
setools==4.6.0
setuptools==82.0.0
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.10.2
soupsieve==2.8
systemd-python==235
tabula-py==2.10.0
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
xkbregistry==0.3
